requirements.txt
asgiref==3.3.1
beautifulsoup4==4.9.3
Django==3.1.3
django-adminlte-3==0.1.6
django-bootstrap4==2.3.1
django-localflavor==3.0.1
psycopg2==2.8.6
python-stdnum==1.14
pytz==2020.4
six==1.11.0
soupsieve==2.0.1
sqlparse==0.4.1
